Moved functions to separate packages and modules
